requirements.txt
nextcord==2.0.0
python-dotenv
requests==2.28.1